requirement.txt
fastapi
uvicorn
dnspython
selenium
transformers
sklearn
joblib
jieba
numpy
pandas

